🏗️ Update project structure
