+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 0.0.4 [2026-01-11]
+## 0.0.5 - 2026-05-24
+
+- Multiple architecture build.
+- Fix code coverage.
+- Updating project plugins, CI and dependencies.
+
+## 0.0.4 - 2026-01-11
 
 - Updating project plugins, CI and dependencies.
