+++ README.md
@@ -26,15 +26,6 @@ Thank you for your continued support!
 
 [![BuyMeACoffee](https://www.buymeacoffee.com/assets/img/guidelines/download-assets-sm-2.svg)](https://www.buymeacoffee.com/edufolly)
 
-## PIX
-
-Sua contribuição ajudará a impulsionar o desenvolvimento de ferramentas de
-qualidade para a comunidade de desenvolvedores Flutter e Dart. Qualquer quantia
-será apreciada.
-Obrigado pelo seu apoio contínuo!
-
-[![PIX](helpers/pix.png)](https://nubank.com.br/pagar/2bt2q/RBr4Szfuwr)
-
 ## Download
 
 ```shell
