+++ pubspec.yaml
@@ -5,7 +5,7 @@ repository: https://github.com/testainers/check
 homepage: https://check.testainers.com
 
 environment:
-  sdk: ^3.9.0
+  sdk: ^3.11.0
 
 dependencies:
   # https://pub.dev/packages/args
@@ -16,13 +16,16 @@ dependencies:
 
 dev_dependencies:
   # https://pub.dev/packages/lints
-  lints: ^6.0.0
+  lints: ^6.1.0
 
   # https://pub.dev/packages/markdown
   markdown: ^7.3.0
 
+  # https://pub.dev/packages/coverage
+  coverage: ^1.15.0
+
   # https://pub.dev/packages/test
-  test: ^1.29.0
+  test: ^1.31.0
 
   # https://pub.dev/packages/testainers
-  testainers: ^0.3.0
+  testainers: ^0.4.0
